Add files via upload
I joined adding to log and counting how many entries in log into one logSystem module. There is still small bug to be fixed. The code notice that there is more entries than it is needed and it does translate those entries to jsObject (Maybe I will need to stringify it, it will be JSON array) and push it to "imaginary"(for now) reporting subsystem function. If it was successfully pushed, logSystem receives "success", but still does not want to clean files.